relocated ParagraphPreview
