+++ Decsys/ClientApp/src/app/screens/admin/EditorScreen/PureEditorScreen.js
@@ -6,9 +6,7 @@ import EditorPageList from "../../../components/EditorPageList";
 import { EmptyState } from "../../../components/ui";
 import { FileAlt } from "styled-icons/fa-solid";
 import { Box, colorVariant } from "@smooth-ui/core-sc";
-import ComponentRender from "../../../components/ComponentRender";
 import ComponentEditor from "../../../components/ComponentEditor";
-import ParagraphPreview from "../../../components/ComponentEditor/ParagraphPreview";
 import ImageUpload from "../../../components/ComponentEditor/ImageUpload";
 import { getComponent } from "../../../utils/component-utils";
 import SurveyPageBody from "../../../components/SurveyPage/Body";
